Quit using the f*cking ObjectProperty kluge and implement the much more succint subscriber collection.
Why? ObjectProperty will accept invalidation listeners, but does not appear to keep them. A quick peek at the mutable-state kangaroo code mess that is ExpressionHelper made me assume that that was the case, but I was too lazy to figure out why getValue was being called on the property and ignored before attaching the subscriber (WTF??????????????????????????) to look any further.